add pymafx_49 kps convention
