+++ mmhuman3d/core/conventions/keypoints_mapping/pymafx_smplx.py
@@ -1,0 +1,27 @@
+"""PyMAF-X in smplx convention."""
+PYMAFX_SMPLX_KEYPOINTS = [
+    'right_ankle',
+    'right_knee',
+    'right_hip_extra',  # LSP
+    'left_hip_extra',  # LSP
+    'left_knee',
+    'left_ankle',
+    'right_wrist',
+    'right_elbow',
+    'right_shoulder',
+    'left_shoulder',
+    'left_elbow',
+    'left_wrist',
+    'neck_extra',  # LSP
+    'headtop',  # LSP
+    'pelvis_extra',  # MPII
+    'thorax_extra',  # MPII
+    'spine_extra',  # H36M
+    'jaw_extra',  # H36M
+    'head_extra',  # H36M
+    'nose',
+    'left_eye',
+    'right_eye',
+    'left_ear',
+    'right_ear',
+]
