+++ mmhuman3d/models/body_models/smplx.py
@@ -1,37 +1,19 @@
 # yapf: disable
-import os
-from dataclasses import dataclass
 from typing import Optional
 
 import numpy as np
 import torch
-import torch.nn as nn
 from smplx import SMPLX as _SMPLX
 from smplx import SMPLXLayer as _SMPLXLayer
-from smplx.body_models import SMPLXOutput
-from smplx.lbs import batch_rodrigues, blend_shapes, vertices2joints
+from smplx.lbs import vertices2joints
 
 from mmhuman3d.core.conventions.keypoints_mapping import (
     convert_kps,
     get_keypoint_num,
 )
-from mmhuman3d.core.conventions.keypoints_mapping.mano import (
-    MANO_LEFT_REORDER_KEYPOINTS,
-    MANO_RIGHT_REORDER_KEYPOINTS,
-)
-from mmhuman3d.core.conventions.keypoints_mapping.openpose import (
-    OPENPOSE_25_KEYPOINTS,
-)
-from mmhuman3d.core.conventions.keypoints_mapping.smplx import SMPLX_KEYPOINTS
-from mmhuman3d.core.conventions.keypoints_mapping.spin_smplx import (
-    SPIN_SMPLX_KEYPOINTS,
-)
 from mmhuman3d.core.conventions.segmentation import body_segmentation
 
 # yapf: enable
-JOINT_NAMES = OPENPOSE_25_KEYPOINTS + SPIN_SMPLX_KEYPOINTS
-FOOT_NAMES = ['bigtoe', 'smalltoe', 'heel']
-SMPLX_JOINT_IDS = {SMPLX_KEYPOINTS[i]: i for i in range(len(SMPLX_KEYPOINTS))}
 
 
 class SMPLX(_SMPLX):
@@ -393,329 +375,3 @@ class SMPLXLayer(_SMPLXLayer):
         if return_full_pose:
             output['full_pose'] = smplx_output.full_pose
         return output
-
-
-@dataclass
-class ModelOutput(SMPLXOutput):
-    """The name of the model output."""
-    smpl_joints: Optional[torch.Tensor] = None
-    joints_J19: Optional[torch.Tensor] = None
-    smplx_vertices: Optional[torch.Tensor] = None
-    flame_vertices: Optional[torch.Tensor] = None
-    lhand_vertices: Optional[torch.Tensor] = None
-    rhand_vertices: Optional[torch.Tensor] = None
-    lhand_joints: Optional[torch.Tensor] = None
-    rhand_joints: Optional[torch.Tensor] = None
-    face_joints: Optional[torch.Tensor] = None
-    lfoot_joints: Optional[torch.Tensor] = None
-    rfoot_joints: Optional[torch.Tensor] = None
-
-
-class SMPLX_ALL(nn.Module):
-    """Extension of the official SMPLX implementation to support more
-    joints."""
-    SMPLX2SMPL_J45 = [i for i in range(22)
-                      ] + [30, 45] + [i for i in range(55, 55 + 21)]
-    JOINT_MAP = {
-        'nose_openpose': 24,
-        'neck_openpose': 12,
-        'right_shoulder_openpose': 17,
-        'right_elbow_openpose': 19,
-        'right_wrist_openpose': 21,
-        'left_shoulder_openpose': 16,
-        'left_elbow_openpose': 18,
-        'left_wrist_openpose': 20,
-        'pelvis_openpose': 0,
-        'right_hip_openpose': 2,
-        'right_knee_openpose': 5,
-        'right_ankle_openpose': 8,
-        'left_hip_openpose': 1,
-        'left_knee_openpose': 4,
-        'left_ankle_openpose': 7,
-        'right_eye_openpose': 25,
-        'left_eye_openpose': 26,
-        'right_ear_openpose': 27,
-        'left_ear_openpose': 28,
-        'left_bigtoe_openpose': 29,
-        'left_smalltoe_openpose': 30,
-        'left_heel_openpose': 31,
-        'right_bigtoe_openpose': 32,
-        'right_smalltoe_openpose': 33,
-        'right_heel_openpose': 34,
-        'right_ankle': 8,
-        'right_knee': 5,
-        'right_hip': 45,
-        'left_hip': 46,
-        'left_knee': 4,
-        'left_ankle': 7,
-        'right_wrist': 21,
-        'right_elbow': 19,
-        'right_shoulder': 17,
-        'left_shoulder': 16,
-        'left_elbow': 18,
-        'left_wrist': 20,
-        'neck': 47,
-        'head_top': 48,
-        'pelvis': 49,
-        'thorax': 50,
-        'spine': 51,
-        'h36m_jaw': 52,
-        'h36m_head': 53,
-        'nose': 24,
-        'left_eye': 26,
-        'right_eye': 25,
-        'left_ear': 28,
-        'right_ear': 27
-    }
-
-    def __init__(self,
-                 batch_size=1,
-                 use_face_contour=True,
-                 gender='neutral',
-                 joint_regressor_train_extra=None,
-                 smplx_model_dir=None,
-                 **kwargs):
-        super().__init__()
-        self.use_face_contour = use_face_contour
-        if gender == 'all':
-            self.genders = ['male', 'female', 'neutral']
-        else:
-            self.genders = [gender]
-        for gender in self.genders:
-            assert gender in ['male', 'female', 'neutral']
-        self.model_dict = nn.ModuleDict({
-            gender: SMPLXLayer(
-                smplx_model_dir,
-                gender=gender,
-                ext='npz',
-                num_betas=10,
-                use_pca=False,
-                batch_size=batch_size,
-                use_face_contour=use_face_contour,
-                num_pca_comps=45,
-                keypoint_src='smplx',
-                keypoint_dst='smplx',
-                **kwargs)
-            for gender in self.genders
-        })
-        self.model_neutral = self.model_dict['neutral']
-        joints = [self.JOINT_MAP[i] for i in JOINT_NAMES]
-        J_regressor_extra = np.load(joint_regressor_train_extra)
-        self.register_buffer(
-            'J_regressor_extra',
-            torch.tensor(J_regressor_extra, dtype=torch.float32))
-        self.joint_map = torch.tensor(joints, dtype=torch.long)
-        smplx_to_smpl = dict(
-            np.load(os.path.join(smplx_model_dir, 'smplx_to_smpl.npz')))
-        self.register_buffer(
-            'smplx2smpl',
-            torch.tensor(smplx_to_smpl['matrix'][None], dtype=torch.float32))
-        smpl2limb_vert_faces = get_partial_smpl()
-        self.smpl2lhand = torch.from_numpy(
-            smpl2limb_vert_faces['lhand']['vids']).long()
-        self.smpl2rhand = torch.from_numpy(
-            smpl2limb_vert_faces['rhand']['vids']).long()
-
-        # left and right hand joint mapping
-        smplx2lhand_joints = [
-            SMPLX_JOINT_IDS[name] for name in MANO_LEFT_REORDER_KEYPOINTS
-        ]
-        smplx2rhand_joints = [
-            SMPLX_JOINT_IDS[name] for name in MANO_RIGHT_REORDER_KEYPOINTS
-        ]
-        self.smplx2lh_joint_map = torch.tensor(
-            smplx2lhand_joints, dtype=torch.long)
-        self.smplx2rh_joint_map = torch.tensor(
-            smplx2rhand_joints, dtype=torch.long)
-
-        # left and right foot joint mapping
-        smplx2lfoot_joints = [
-            SMPLX_JOINT_IDS['left_{}'.format(name)] for name in FOOT_NAMES
-        ]
-        smplx2rfoot_joints = [
-            SMPLX_JOINT_IDS['right_{}'.format(name)] for name in FOOT_NAMES
-        ]
-        self.smplx2lf_joint_map = torch.tensor(
-            smplx2lfoot_joints, dtype=torch.long)
-        self.smplx2rf_joint_map = torch.tensor(
-            smplx2rfoot_joints, dtype=torch.long)
-
-        for g in self.genders:
-            J_template = torch.einsum('ji,ik->jk', [
-                self.model_dict[g].J_regressor[:24],
-                self.model_dict[g].v_template
-            ])
-            J_dirs = torch.einsum('ji,ikl->jkl', [
-                self.model_dict[g].J_regressor[:24],
-                self.model_dict[g].shapedirs
-            ])
-
-            self.register_buffer(f'{g}_J_template', J_template)
-            self.register_buffer(f'{g}_J_dirs', J_dirs)
-
-    def forward(self, *args, **kwargs):
-        """Forward function."""
-        batch_size = kwargs['body_pose'].shape[0]
-        kwargs['get_skin'] = True
-        if 'pose2rot' not in kwargs:
-            kwargs['pose2rot'] = True
-        if 'gender' not in kwargs:
-            kwargs['gender'] = 2 * torch.ones(batch_size).to(
-                kwargs['body_pose'].device)
-
-        # pose for 55 joints: 1, 21, 15, 15, 1, 1, 1
-        pose_keys = [
-            'global_orient', 'body_pose', 'left_hand_pose', 'right_hand_pose',
-            'jaw_pose', 'leye_pose', 'reye_pose'
-        ]
-        param_keys = ['betas'] + pose_keys
-        if kwargs['pose2rot']:
-            for key in pose_keys:
-                if key in kwargs:
-                    kwargs[key] = batch_rodrigues(
-                        kwargs[key].contiguous().view(-1, 3)).view(
-                            [batch_size, -1, 3, 3])
-        if kwargs['body_pose'].shape[1] == 23:
-            # remove hand pose in the body_pose
-            kwargs['body_pose'] = kwargs['body_pose'][:, :21]
-        gender_idx_list = []
-        smplx_vertices, smplx_joints = [], []
-        for gi, g in enumerate(['male', 'female', 'neutral']):
-            gender_idx = ((kwargs['gender'] == gi).nonzero(as_tuple=True)[0])
-            if len(gender_idx) == 0:
-                continue
-            gender_idx_list.extend([int(idx) for idx in gender_idx])
-            gender_kwargs = {
-                'get_skin': kwargs['get_skin'],
-                'pose2rot': kwargs['pose2rot']
-            }
-            gender_kwargs.update(
-                {k: kwargs[k][gender_idx]
-                 for k in param_keys if k in kwargs})
-            gender_smplx_output = self.model_dict[g].forward(
-                *args, **gender_kwargs)
-            smplx_vertices.append(gender_smplx_output['vertices'])
-            smplx_joints.append(gender_smplx_output['joints'])
-
-        idx_rearrange = [
-            gender_idx_list.index(i)
-            for i in range(len(list(gender_idx_list)))
-        ]
-        idx_rearrange = torch.tensor(idx_rearrange).long().to(
-            kwargs['body_pose'].device)
-
-        smplx_vertices = torch.cat(smplx_vertices)[idx_rearrange]
-        smplx_joints = torch.cat(smplx_joints)[idx_rearrange]
-
-        lhand_joints = smplx_joints[:, self.smplx2lh_joint_map]
-        rhand_joints = smplx_joints[:, self.smplx2rh_joint_map]
-        face_joints = smplx_joints[:, -68:] if self.use_face_contour \
-            else smplx_joints[:, -51:]
-        lfoot_joints = smplx_joints[:, self.smplx2lf_joint_map]
-        rfoot_joints = smplx_joints[:, self.smplx2rf_joint_map]
-
-        smpl_vertices = torch.bmm(
-            self.smplx2smpl.expand(batch_size, -1, -1), smplx_vertices)
-        lhand_vertices = smpl_vertices[:, self.smpl2lhand]
-        rhand_vertices = smpl_vertices[:, self.smpl2rhand]
-        extra_joints = vertices2joints(self.J_regressor_extra, smpl_vertices)
-        # smpl_output.joints: [B, 45, 3]  extra_joints: [B, 9, 3]
-        smplx_j45 = smplx_joints[:, self.SMPLX2SMPL_J45]
-        joints = torch.cat([smplx_j45, extra_joints], dim=1)
-        smpl_joints = smplx_j45[:, :24]
-        joints = joints[:, self.joint_map, :]  # [B, 49, 3]
-        joints_J24 = joints[:, -24:, :]
-        # Indices to get the 14 LSP joints from the ground truth joints
-        J24_TO_J17 = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 18, 14, 16, 17]
-        J24_TO_J19 = J24_TO_J17[:14] + [19, 20, 21, 22, 23]
-        joints_J19 = joints_J24[:, J24_TO_J19, :]
-        output = ModelOutput(
-            vertices=smpl_vertices,
-            smplx_vertices=smplx_vertices,
-            lhand_vertices=lhand_vertices,
-            rhand_vertices=rhand_vertices,
-            joints=joints,
-            joints_J19=joints_J19,
-            smpl_joints=smpl_joints,
-            lhand_joints=lhand_joints,
-            rhand_joints=rhand_joints,
-            lfoot_joints=lfoot_joints,
-            rfoot_joints=rfoot_joints,
-            face_joints=face_joints,
-        )
-        return output
-
-    def get_tpose(self, betas=None, gender=None):
-        """Get tpose joints.
-
-        Args:
-            betas (betas, optional): Defaults to None.
-            gender (str, optional): Defaults to None.
-
-        Returns:
-            smplx_joints (torch.Tensor): smplx joints in shape [1, n_joints, 3]
-        """
-        kwargs = {}
-        if betas is None:
-            betas = torch.zeros(1, 10).to(self.J_regressor_extra.device)
-        kwargs['betas'] = betas
-
-        batch_size = kwargs['betas'].shape[0]
-        device = kwargs['betas'].device
-
-        if gender is None:
-            kwargs['gender'] = 2 * torch.ones(batch_size).to(device)
-        else:
-            kwargs['gender'] = gender
-
-        param_keys = ['betas']
-
-        gender_idx_list = []
-        smplx_joints = []
-        for gi, g in enumerate(['male', 'female', 'neutral']):
-            gender_idx = ((kwargs['gender'] == gi).nonzero(as_tuple=True)[0])
-            if len(gender_idx) == 0:
-                continue
-            gender_idx_list.extend([int(idx) for idx in gender_idx])
-            gender_kwargs = {}
-            gender_kwargs.update(
-                {k: kwargs[k][gender_idx]
-                 for k in param_keys if k in kwargs})
-
-            J = getattr(self, f'{g}_J_template').unsqueeze(0) + blend_shapes(
-                gender_kwargs['betas'], getattr(self, f'{g}_J_dirs'))
-
-            smplx_joints.append(J)
-
-        idx_rearrange = [
-            gender_idx_list.index(i)
-            for i in range(len(list(gender_idx_list)))
-        ]
-        idx_rearrange = torch.tensor(idx_rearrange).long().to(device)
-
-        smplx_joints = torch.cat(smplx_joints)[idx_rearrange]
-        return smplx_joints
-
-
-def get_partial_smpl():
-    """Get partial mesh of SMPL.
-
-    Returns:
-        part_vert_faces
-    """
-    part_vert_faces = {}
-
-    for part in [
-            'lhand', 'rhand', 'face', 'arm', 'forearm', 'larm', 'rarm',
-            'lwrist', 'rwrist'
-    ]:
-        part_vid_fname = f'data/partial_mesh/smpl_{part}_vids.npz'
-        if os.path.exists(part_vid_fname):
-            part_vids = np.load(part_vid_fname)
-            part_vert_faces[part] = {
-                'vids': part_vids['vids'],
-                'faces': part_vids['faces']
-            }
-        else:
-            raise FileNotFoundError(f'{part_vid_fname} does not exist!')
-    return part_vert_faces
