+++ tests/test_pymafx_pipeline.py
@@ -11,7 +11,7 @@ from torch.utils.data import DataLoader
 
 from mmhuman3d.apis import init_model
 from mmhuman3d.data.datasets import build_dataset
-from mmhuman3d.utils.keypoint_utils import process_kps2d
+from mmhuman3d.utils.keypoint_utils import transform_kps2d
 
 
 def _init_openpifpaf(parser):
@@ -68,7 +68,7 @@ def _setup_parser():
 
 def test_process_kps2d():
     pred_hand_proj = torch.ones((1, 390, 2))
-    kps2d = process_kps2d(pred_hand_proj, torch.ones((1, 2, 3)))
+    kps2d = transform_kps2d(pred_hand_proj, torch.ones((1, 2, 3)))
     assert kps2d.shape == (1, 390, 2)
